+++ src/NeatMapper/Matchers/IdentityMatcher.cs
@@ -1,0 +1,50 @@
+﻿using System;
+
+namespace NeatMapper {
+	/// <summary>
+	/// Singleton <see cref="IMatcher"/> which matches the same object (by value/reference with the equality operator).
+	/// Supports only the same source/destination types. Can be used to merge collections of elements of the same type.
+	/// </summary>
+	public sealed class IdentityMatcher : IMatcher, IMatcherFactory {
+		/// <summary>
+		/// Singleton instance of the matcher.
+		/// </summary>
+		public static readonly IMatcher Instance = new IdentityMatcher();
+
+
+		private IdentityMatcher() { }
+
+
+		public bool CanMatch(Type sourceType, Type destinationType, MappingOptions? mappingOptions = null) {
+			if (sourceType == null)
+				throw new ArgumentNullException(nameof(sourceType));
+			if (destinationType == null)
+				throw new ArgumentNullException(nameof(destinationType));
+
+			return sourceType == destinationType;
+		}
+
+		public bool Match(object? source, Type sourceType, object? destination, Type destinationType, MappingOptions? mappingOptions = null) {
+			if (!CanMatch(sourceType, destinationType) || sourceType.IsGenericTypeDefinition || destinationType.IsGenericTypeDefinition)
+				throw new MapNotFoundException((sourceType, destinationType));
+
+			TypeUtils.CheckObjectType(source, sourceType, nameof(source));
+			TypeUtils.CheckObjectType(destination, destinationType, nameof(destination));
+
+			return source == destination;
+		}
+
+		public IMatchMapFactory MatchFactory(Type sourceType, Type destinationType, MappingOptions? mappingOptions = null) {
+			if (!CanMatch(sourceType, destinationType) || sourceType.IsGenericTypeDefinition || destinationType.IsGenericTypeDefinition)
+				throw new MapNotFoundException((sourceType, destinationType));
+
+			return new DefaultMatchMapFactory(sourceType, destinationType,
+				(source, destination) => {
+					TypeUtils.CheckObjectType(source, sourceType, nameof(source));
+					TypeUtils.CheckObjectType(destination, destinationType, nameof(destination));
+
+					return source == destination;
+				});
+		}
+	}
+}